+++ CHANGES.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `device='auto'` to select CUDA when available and CPU otherwise.
+- Add `device='auto'` to select hardware acceleration like CUDA when
+  available, and CPU otherwise.
 
 ### Changed